Headers for device related files
